vim-git: add syntax highlighting for diffs in gitsendemail
This patch adds support for --annotate option of git-send-email in a
number of ways. First, it adds the ability to detect git-format-patch
style files, by checking if the first line matches the format generated
by git-format-patch. Second, it adds support for highlighting the diff
portion using a region and include. This enables correct textwidth
settings for the commit message, as well as proper diff highlighting and
indentation settings for the rest of the patch contents.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>