Fix highlighting "git log --parents"
When passing the "--parents" options to "git log" two commit hashes are
shown in the commit header in the log; since this case is not covered by
vim syntax/git.vim, the highlighting does not work.

Fix this by matching the commit hash, and the leading space, *at least*
once instead of *exactly* once.